beforeInstall if condition

+++ codedeploy/beforeInstall.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
 #Clean up tomcat artifacts
-sudo rm /opt/tomcat8/webapps/ROOT.war
-sudo rm /opt/tomcat8/webapps/ROOT
+if [ -e /opt/tomcat8/webapps/ROOT.war ]
+then
+    sudo rm /opt/tomcat8/webapps/ROOT.war
+elif [-e /opt/tomcat8/webapps/ROOT]
+then
+    sudo rm /opt/tomcat8/webapps/ROOT
+else
+    echo "No ROOT/ROOT.war exists"
+fi
+
+
